refact: Create separate directory for ChatMessages component. Break chat component in separate components.
